Implement portal for modal display and refactor game completion handling
