+++ src/pages/GamePage.tsx
@@ -3,9 +3,33 @@ import { GameControls } from "../components/game/GameControls";
 import { HanoiGame } from "../components/game/HanoiGame";
 import type { GamePageProps } from "../types/ui.types";
 import { useGame } from "../hooks/useGame";
-import { useEffect } from "react";
+import { useState } from "react";
+import Portal from "../components/common/Portal";
+import type { GameStatistic } from "../types/game.types";
+import { ResultsPage } from "./ResultsPage";
 
-export function GamePage({ difficulty, onQuit, onGameComplete }: GamePageProps) {
+export function GamePage({ difficulty, onQuit }: GamePageProps) {
+
+    const handleGameComplete = (movesCount: number, timePassed: number, timeRemaining: number | null, isGameWon: boolean) => {
+        const minMoves = Math.pow(2, difficulty.disks) - 1;
+        const efficiency = Math.round((minMoves / movesCount) * 100);
+
+        setGameStats({
+            movesCount,
+            timePassed,
+            timeRemaining,
+            difficulty,
+            minMoves,
+            efficiency,
+            isGameWon
+        });
+        setShowModal(true);
+    }
+
+    const handlePlayAgain = () => {
+        resetGame();
+        setShowModal(false);
+    }
 
     const {
         gameState,
@@ -16,15 +40,14 @@ export function GamePage({ difficulty, onQuit, onGameComplete }: GamePageProps) 
         resetGame,
         pauseGame,
         resumeGame,
-    } = useGame(difficulty, onGameComplete);
+    } = useGame(difficulty, handleGameComplete);
 
-    useEffect(() => {
-        console.log(gameState.towers)
-    }, [gameState.towers]);
+    const [showModal, setShowModal] = useState(false);
+    const [gameStats, setGameStats] = useState<GameStatistic | null>(null);
 
     return (
-        <Layout>         
-            <GameControls 
+        <Layout>
+            <GameControls
                 movesCount={gameState.movesCount}
                 timePassed={timePassed}
                 timeRemaining={timeRemaining}
@@ -32,10 +55,20 @@ export function GamePage({ difficulty, onQuit, onGameComplete }: GamePageProps) 
                 onReset={resetGame}
                 onPause={pauseGame}
                 onResume={resumeGame}
-                onQuite={onQuit} 
+                onQuite={onQuit}
             />
 
             <HanoiGame gameState={gameState} onTowerSelect={handleTowerSelect} />
+
+            {showModal &&
+                <Portal>
+                    <ResultsPage
+                        gameStatistic={gameStats!}
+                        onPlayAgain={handlePlayAgain}
+                        onMainMenu={onQuit}
+                    />
+                </Portal>
+            }
         </Layout>
     )
 }
